// src/types/adminType.ts
export type UserStatus = 'active' | 'banned' | 'deleted';
export type AdminTeacherRequestStatus = 'pending' | 'approved' | 'rejected';
export type MaterialType = 'file' | 'slide' | '2d_render' | 'quiz';
export type MaterialStatus = 'draft' | 'published' | 'reviewed' | 'deleted';

export interface AdminUser {
  _id: string;
  username: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  date_create: string;
  teacher?: {
    _id: string;
    user_id: string;
    credential?: string;
    fileName?: string;
  } | null;
  admin?: {
    _id: string;
    user_id: string;
    authorization_lvl: 1 | 2;
  } | null;
}

export interface CreateUserRequest {
  username: string;
  email: string;
  password: string;
  role: UserRole;
  credentialFile?: File | null;
}

export interface UpdateUserRequest {
  username?: string;
  email?: string;
}